Handle account state msgs
