+++ modules/token/handle_msg.go
@@ -1,12 +1,10 @@
 package token
 
 import (
-	"encoding/base64"
 	"fmt"
 
 	"github.com/forbole/soljuno/client"
 	"github.com/forbole/soljuno/db"
-	"github.com/forbole/soljuno/solana/account"
 	"github.com/forbole/soljuno/solana/program/token"
 	"github.com/forbole/soljuno/types"
 	"github.com/rs/zerolog/log"
@@ -33,6 +31,7 @@ func HandleMsg(msg types.Message, tx types.Tx, db db.TokenDb, client client.Prox
 	case "transferChecked":
 		return nil
 
+	// Delegate msgs
 	case "approve":
 		return handleMsgApprove(msg, db, client)
 	case "approveChecked":
@@ -40,20 +39,26 @@ func HandleMsg(msg types.Message, tx types.Tx, db db.TokenDb, client client.Prox
 	case "revoke":
 		return handleMsgRevoke(msg, db, client)
 
+	case "setAuthority":
+		return nil
+
+	// Token supply msgs
 	case "mintTo":
 		return handleMsgMintTo(msg, db, client)
 	case "mintToChecked":
 		return handleMsgMintToChecked(msg, db, client)
-
 	case "burn":
 		return handleMsgBurn(msg, db, client)
 	case "burnChecked":
 		return handleMsgBurnChecked(msg, db, client)
 
+	// Account state msgs
 	case "closeAccount":
+		return handleMsgCloseAccount(msg, db, client)
 	case "freezeAccount":
+		return handleMsgFreezeAccount(msg, db, client)
 	case "thawAccount":
-		return nil
+		return handleMsgThawAccount(msg, db, client)
 	}
 
 	log.Info().Str("module", "token").Str("message", msg.Value.Type()).Uint64("slot", tx.Slot).
@@ -117,34 +122,6 @@ func handleMsgInitializeMultisig(msg types.Message, tx types.Tx, db db.TokenDb) 
 	return nil
 }
 
-// updateDelegation properly stores the statement of delegation inside the database
-func updateDelegation(source string, db db.TokenDb, client client.Proxy) error {
-	info, err := client.AccountInfo(source)
-	if err != nil {
-		return err
-	}
-
-	if info.Value == nil {
-		return db.SaveDelegate(source, "", info.Context.Slot, 0)
-	}
-
-	bz, err := base64.StdEncoding.DecodeString(info.Value.Data[0])
-	if err != nil {
-		return err
-	}
-
-	tokenAccount, ok := account.Parse(token.ProgramID, bz).(account.TokenAccount)
-	if !ok {
-		return db.SaveDelegate(source, "", info.Context.Slot, 0)
-	}
-
-	if !tokenAccount.Delegate.Option.Bool() {
-		return db.SaveDelegate(source, "", info.Context.Slot, 0)
-	}
-
-	return db.SaveDelegate(source, tokenAccount.Delegate.Value.String(), info.Context.Slot, tokenAccount.DelegateAmount)
-}
-
 // handleMsgApproveChecked handles a MsgApprove
 func handleMsgApprove(msg types.Message, db db.TokenDb, client client.Proxy) error {
 	instruction, ok := msg.Value.Data().(token.ParsedApprove)
@@ -170,25 +147,6 @@ func handleMsgRevoke(msg types.Message, db db.TokenDb, client client.Proxy) erro
 		return fmt.Errorf("instruction does not match approveChecked type: %s", msg.Value.Type())
 	}
 	return updateDelegation(instruction.Source, db, client)
-}
-
-// updateTokenSupply properly stores the supply of the given mint inside the database
-func updateTokenSupply(mint string, db db.TokenDb, client client.Proxy) error {
-	info, err := client.AccountInfo(mint)
-	if err != nil {
-		return err
-	}
-
-	bz, err := base64.StdEncoding.DecodeString(info.Value.Data[0])
-	if err != nil {
-		return err
-	}
-
-	token, ok := account.Parse(token.ProgramID, bz).(account.TokenMint)
-	if !ok {
-		return fmt.Errorf("failed to parse token:%s", mint)
-	}
-	return db.SaveTokenSupply(mint, info.Context.Slot, token.Supply)
 }
 
 // handleMsgMintTo handles a MsgMintTo
@@ -225,4 +183,31 @@ func handleMsgBurnChecked(msg types.Message, db db.TokenDb, client client.Proxy)
 		return fmt.Errorf("instruction does not match burnChecked type: %s", msg.Value.Type())
 	}
 	return updateTokenSupply(instruction.Mint, db, client)
+}
+
+// handleMsgCloseAccount handles a MsgCloseAccount
+func handleMsgCloseAccount(msg types.Message, db db.TokenDb, client client.Proxy) error {
+	instruction, ok := msg.Value.Data().(token.ParsedCloseAccount)
+	if !ok {
+		return fmt.Errorf("instruction does not match closeAccount type: %s", msg.Value.Type())
+	}
+	return updateAccountState(instruction.Account, db, client)
+}
+
+// handleMsgFreezeAccount handles a MsgFreezeAccount
+func handleMsgFreezeAccount(msg types.Message, db db.TokenDb, client client.Proxy) error {
+	instruction, ok := msg.Value.Data().(token.ParsedFreezeAccount)
+	if !ok {
+		return fmt.Errorf("instruction does not match freezeAccount type: %s", msg.Value.Type())
+	}
+	return updateAccountState(instruction.Account, db, client)
+}
+
+// handleMsgThawAccount handles a MsgThawAccount
+func handleMsgThawAccount(msg types.Message, db db.TokenDb, client client.Proxy) error {
+	instruction, ok := msg.Value.Data().(token.ParsedThawAccount)
+	if !ok {
+		return fmt.Errorf("instruction does not match thawAccount type: %s", msg.Value.Type())
+	}
+	return updateAccountState(instruction.Account, db, client)
 }
